Lightmap generator raycaster first working version

+++ jni/raycaster/triangle.h
@@ -16,9 +16,9 @@ public:
     glm::vec3 a;
     glm::vec3 b;
     glm::vec3 c;
-    glm::ivec2 aID;
-    glm::ivec2 bID;
-    glm::ivec2 cID;
+    glm::vec2 aID;
+    glm::vec2 bID;
+    glm::vec2 cID;
     int lmIndex;
     int tIndex;
     long lastTestID;
